requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Session
itsdangerous==2.2.0

# Scheduler
apscheduler==3.10.4

# System metrics
psutil==6.1.0

# Templates
jinja2==3.1.4

# HTTP client for cronjob execution
httpx==0.28.1

# Config & env
python-dotenv==1.0.1
